refactor: 상수 및 도메인 분리
