+++ src/data/constants.ts
@@ -1,0 +1,13 @@
+export const CATEGORY_LIST = {
+  all: "전체",
+  korean: "한식",
+  japanese: "일식",
+  western: "양식",
+  asian: "아시안",
+  etc: "기타",
+};
+
+export const SORT_LIST = {
+  name: "이름순",
+  distance: "거리순",
+};
